Asset API development completed On branch main Changes to be committed: 	new file:   src/main/java/com/adithya/trading/controller/AssetController.java 	new file:   src/main/java/com/adithya/trading/model/Asset.java 	new file:   src/main/java/com/adithya/trading/repository/AssetRepository.java 	new file:   src/main/java/com/adithya/trading/service/AssetService.java 	new file:   src/main/java/com/adithya/trading/service/AssetServiceImpl.java 	modified:   src/main/java/com/adithya/trading/service/OrderServiceImpl.java

+++ src/main/java/com/adithya/trading/service/OrderServiceImpl.java
@@ -10,6 +10,7 @@ import org.springframework.transaction.annotation.Transactional;
 
 import com.adithya.trading.domain.OrderStatus;
 import com.adithya.trading.domain.OrderType;
+import com.adithya.trading.model.Asset;
 import com.adithya.trading.model.Coin;
 import com.adithya.trading.model.Order;
 import com.adithya.trading.model.OrderItem;
@@ -27,6 +28,9 @@ public class OrderServiceImpl implements OrderService {
 
     @Autowired
     private OrderItemRepository orderItemRepository;
+
+    @Autowired
+    private AssetService assetService;
 
     @Override
     public Order createOrder(User user, OrderItem orderItem, OrderType orderType) {
@@ -75,7 +79,13 @@ public class OrderServiceImpl implements OrderService {
         order.setStatus(OrderStatus.SUCCESS);
         order.setOrderType(OrderType.BUY);
         Order savedOrder = orderRepository.save(order);
-        
+        Asset oldAsset = assetService.findAssetByUserIdAndCoinId(order.getUser().getId(), order.getOrderItem().getCoin().getId());
+        if(oldAsset == null) {
+            assetService.createAsset(user, orderItem.getCoin(), orderItem.getQuantity());
+        } else {
+            assetService.updateAsset(oldAsset.getId(), quantity)
+        }
+
         return savedOrder;
     }
 
@@ -85,25 +95,28 @@ public class OrderServiceImpl implements OrderService {
             throw new Exception("Quantity should be >0");
         }
         double sellPrice = coin.getCurrentPrice();
-        double buyPrice = assetToSell.getPrice();
-        OrderItem orderItem = createOrderItem(coin, quantity, buyPrice, sellPrice);
-        Order order = createOrder(user, orderItem, OrderType.SELL);
-        orderItem.setOrder(order);
-        
-        if(assetToSell.getQuantity() >= quantity) {
-            order.setStatus(OrderStatus.SUCCESS);
-            order.setOrderType(OrderType.SELL);
-            Order savedOrder = orderRepository.save(order);
+        Asset assetToSell = assetService.findAssetByUserIdAndCoinId(user.getId(), coin.getId());
+        double buyPrice = assetToSell.getBuyPrice();
+        if(assetToSell != null) {
+            OrderItem orderItem = createOrderItem(coin, quantity, buyPrice, sellPrice);
+            Order order = createOrder(user, orderItem, OrderType.SELL);
+            orderItem.setOrder(order);
+            
+            if(assetToSell.getQuantity() >= quantity) {
+                order.setStatus(OrderStatus.SUCCESS);
+                order.setOrderType(OrderType.SELL);
+                Order savedOrder = orderRepository.save(order);
 
-            walletService.payOrderPayment(order, user);
-
-            Asset updatedAsset = assetService.updatedAsset(assetToSell.getId(), -quantity);
-            if(updatedAsset.getQuantity()*coin.getCurrentPrice() <= 1) {
-                assetService.deleteAsset(updatedAsset.getId());
-            }
-            return savedOrder;
+                walletService.payOrderPayment(order, user);
+                Asset updatedAsset = assetService.updateAsset(assetToSell.getId(), -quantity);
+                if(updatedAsset.getQuantity()*coin.getCurrentPrice() <= 1) {
+                    assetService.deleteAsset(updatedAsset.getId());
+                }
+                return savedOrder;
+            }  
+            throw new Exception("Insufficient Quantity to Sell");
         }
-        throw new Exception("Insufficient Quantity to Sell");
+        throw new Exception("Asset not found");
     }
 
     @Override
